refactor: move AbstractItemContainer to container package
